fix: database operation on main thread
fix: the bug of database operation on main thread

bug was introduce in 1fbcc9b892631d1c1797fa1f51088e9358efd327

This commit is not stand alone all previous commit are part of it. after bug are part of it.

be very careful revert any of this commit as standalone

chore(depsfix): fix a issue in build.gradle

refactor: ViewPassword.kt to use coroutines

refactor: UpdatePassword.kt to use coroutines

refactor: SavePassword.kt to use coroutines

refactor: LoadPassword.kt to use coroutines

refactor: update the controller to use suspend function

the commit is not complete as a whole

feat: make every db operation a suspend function and use flows

the commit is not complete as a whole

chore(deps): add dependency for coroutines

chore(deps): improve the build.gradle and make morder